+++ packages/next-common/components/overview/accountInfo/components/multisigManagePrompt.js
@@ -1,19 +1,15 @@
 import Prompt from "./prompt";
 import { PromptTypes } from "next-common/components/scrollPrompt";
 import { CACHE_KEY } from "next-common/utils/constants";
-import { useMemo, useEffect } from "react";
+import { useMemo } from "react";
 import Link from "next/link";
-import { useChain } from "next-common/context/chain";
-import {
-  fetchMyMultisigsCount,
-  fetchMyMultisigs,
-  myMultisigsCountSelector,
-} from "next-common/store/reducers/multisigSlice";
-import { useDispatch, useSelector } from "react-redux";
+import { myMultisigsCountSelector } from "next-common/store/reducers/multisigSlice";
+import { useSelector } from "react-redux";
 import useRealAddress from "next-common/utils/hooks/useRealAddress";
-import getChainSettings from "next-common/utils/consts/settings";
+import { useChainSettings } from "next-common/context/chain";
 import { usePathname } from "next/navigation";
 import { myMultisigsSelector } from "next-common/store/reducers/multisigSlice";
+import useSubscribeMyActiveMultisigs from "next-common/components/overview/accountInfo/hook/useSubscribeMyActiveMultisigs";
 
 const getNeedApprovalCount = (multisigs, address) => {
   const needApprovalItems = multisigs?.filter((item) => {
@@ -37,17 +33,15 @@ function ManageLink({ manageContent }) {
 }
 
 export default function MultisigManagePrompt() {
-  const dispatch = useDispatch();
-  const chain = useChain();
   const realAddress = useRealAddress();
   const myMultisigsCount = useSelector(myMultisigsCountSelector) || 0;
   const myMultisigs = useSelector(myMultisigsSelector);
+  const chainSettings = useChainSettings();
   const { items: multisigs = [], total = 0 } = myMultisigs || {};
   const pathname = usePathname();
-
   const isAccountMultisigPage = pathname.startsWith("/account/multisigs");
 
-  const settings = getChainSettings(chain);
+  useSubscribeMyActiveMultisigs(isAccountMultisigPage);
 
   const needApprovalCount = useMemo(() => {
     if (total === 0) {
@@ -57,18 +51,8 @@ export default function MultisigManagePrompt() {
     return getNeedApprovalCount(multisigs, realAddress);
   }, [multisigs, realAddress, total]);
 
-  useEffect(() => {
-    if (settings?.multisigApiPrefix && realAddress) {
-      dispatch(fetchMyMultisigsCount(chain, realAddress));
-
-      if (!isAccountMultisigPage) {
-        dispatch(fetchMyMultisigs(chain, realAddress));
-      }
-    }
-  }, [dispatch, chain, realAddress, settings, isAccountMultisigPage]);
-
   const promptContent = useMemo(() => {
-    if (!settings?.multisigApiPrefix || myMultisigsCount === 0) {
+    if (!chainSettings?.multisigApiPrefix || myMultisigsCount === 0) {
       return null;
     }
 
@@ -86,7 +70,12 @@ export default function MultisigManagePrompt() {
         {!isAccountMultisigPage && <ManageLink manageContent={manageContent} />}
       </Prompt>
     );
-  }, [myMultisigsCount, needApprovalCount, settings, isAccountMultisigPage]);
+  }, [
+    myMultisigsCount,
+    needApprovalCount,
+    chainSettings,
+    isAccountMultisigPage,
+  ]);
 
   return promptContent;
 }