+++ packages/next-common/context/coretime/sale/index.js
@@ -1,15 +1,24 @@
 import CoretimeActiveSaleSubscriber from "next-common/context/coretime/sale/activeSubscriber";
 import { usePageProps } from "next-common/context/page";
 import { CoretimeSaleProvider } from "next-common/context/coretime/sale/provider";
+import CoretimeDetailSaleSubscriber from "./detailSubscriber";
 
 export function CoretimeActiveSaleProvider({ children }) {
   const { coretimeSale } = usePageProps();
 
   return (
     <CoretimeSaleProvider value={coretimeSale}>
-      <CoretimeActiveSaleSubscriber>
-        {children}
-      </CoretimeActiveSaleSubscriber>
+      <CoretimeActiveSaleSubscriber>{children}</CoretimeActiveSaleSubscriber>
+    </CoretimeSaleProvider>
+  );
+}
+
+export function CoretimeDetailSaleProvider({ children }) {
+  const { coretimeSale } = usePageProps();
+
+  return (
+    <CoretimeSaleProvider value={coretimeSale}>
+      <CoretimeDetailSaleSubscriber>{children}</CoretimeDetailSaleSubscriber>
     </CoretimeSaleProvider>
   );
 }