+++ packages/next-common/context/proxy/index.js
@@ -2,6 +2,7 @@ import React, { useMemo } from "react";
 import useQueryMyProxied from "next-common/hooks/useQueryMyProxies";
 import useAllOnChainProxies from "next-common/hooks/useAllOnChainProxies";
 import { useUser } from "../user";
+import { useChainSettings } from "../chain";
 
 const ProxiesContext = React.createContext(null);
 
@@ -41,6 +42,17 @@ export function OnChainProxiesProvider({ children }) {
     >
       {children}
     </ProxiesContext.Provider>
+  );
+}
+
+export function GeneralProxiesProvider({ children }) {
+  const { modules: { proxy: { provider = "chain" } = {} } = {} } =
+    useChainSettings();
+
+  return provider === "server" ? (
+    <ServerProxiesProvider>{children}</ServerProxiesProvider>
+  ) : (
+    <OnChainProxiesProvider>{children}</OnChainProxiesProvider>
   );
 }